requirements.txt
adhanpy                  # Islamic prayer time calculation (port of batoulapps/Adhan)
tzdata>=2024.1          # required on Windows; no-op on Linux (uses system tz)
pygame>=2.5.0
flask>=3.0.0
waitress>=3.0.0
pydantic>=2.0.0
watchdog>=4.0.0
hijri-converter>=2.3.0
PyYAML>=6.0.0

# Dev / test
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.4.0
mypy>=1.10.0
